c299: extract MetricsTab + RateDistortionCurveSvg + MutualInfoTable + METHOD_COLORS (#441 2.1)
Workspace.tsx 7906 LOC -> 7561 LOC (-345 LOC). Entry chunk 195.85 KB -> 190.20 KB.
18 tabs extracted + lazy. Cumulative since c262: -39%. typecheck clean, 44/44 tests pass.
